+++ app/src/main/java/com/example/repairtool/utilities/navigation/Navigation.kt
@@ -1,2 +1,34 @@
 package com.example.repairtool.utilities.navigation
 
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavType
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navArgument
+import com.example.repairtool.LoginNav
+import com.example.repairtool.volunteer.RepairListScreen
+
+//This function creates the navigation from loginscreen to the repairlist (Seen by volunteers)
+@Composable
+fun LoginNavigation() {
+    val navController = rememberNavController()
+    NavHost(navController = navController, startDestination = Screen.LoginScreen.route) {
+        composable(route = Screen.LoginScreen.route) {
+            LoginNav(navController)
+        }
+        composable(
+            route = Screen.RepairListScreen.route + "/{name}",
+            arguments = listOf(
+                navArgument("name") {
+                    type = NavType.StringType
+                    defaultValue = ""
+                    nullable = true
+                }
+            )
+        ) {
+                entry ->
+            RepairListScreen(name = entry.arguments?.getString("name"))
+        }
+    }
+}
